Updates Vanilla Expanded Framework (#69)
This change patches AnimalBehaviours namespace. Additionally, this change changes the file structure to (hopefully) make it more readable.

It makes Witcher Monster Hunt patch obsolete, as one of the methods it was patching was removed and uses framework for it, and the other is not calling any RNG methods, even though it initializes them. Fixes #68