Added mostly complete compat for Work Tab
The only feature that I wasn't able to sync was the favorites.

Doing it would take some effort, would possibly require syncing a directory, and would also need to make sure the IDs for the favorites are synced - which would require either the dialog to be opened for all players (which could be rather simple), or doing a lot of weird syncing.

So yeah, I decided it was not worth the effort to sync the favorites feature from the mod.