Fix the VFE-Pirates error
Including it as it seems my previous commit changed (I assume) line endings making a big mess. Since it's a minor change relating to it, I decided to not bother making another PR that would either have to wait for the other one to be included, or making identical mess.