Sync camera-dependent stuff like VFE-Medieval archery range. Fix VFE-Pirates error. (#212)
* Sync camera-dependent stuff

Fixes VFE-Medieval archery range. It was spawning motes only if seen by player, and it may not be the case for all players in MP.

* Fix the VFE-Pirates error

Including it as it seems my previous commit changed (I assume) line endings making a big mess. Since it's a minor change relating to it, I decided to not bother making another PR that would either have to wait for the other one to be included, or making identical mess.

Co-authored-by: Sokyran <toCy&#$VIH$1cSOu8s&Fv@UVN%L3snN*H09$1^Xghq*hKo9Itnmq#bNf8Ld67#T$#a&cEaY1>