Reworked dialog closing syncing and pause locks
The new class DialogUtilities handles all the new functionality. It will patch constructor of dialogs of specified type, which will add the open dialog to the list of open dialogs for closing and/or pause locks.

We have a prefix patch to WindowStack.TryRemove, which - if the dialog is inside of the list for close syncing - will stop it from being closed and instead do it in synced context. In case of pausing dialogs it'll remove it from the list of those (if it's also a dialog that sync closing - it'll do it after synced close).

As for pause lock - using a single pause lock for all dialogs should help with performance - as MP will no longer have to iterate over however many pause locks we registered - and instead check if the list of pausing dialogs is empty or not. In case it's locking the pause, it'll also try removing all null or closed dialogs, to prevent situations where the dialog was not properly removed from the list.

This also goes through all mods that had pause locks in place, and modifies them to use this new way of handling those. This change significantly simplifies those patches, as there's no longer need to handle situations where code closing the dialog is inaccessible/hard to access.