Improved MVCF patch, should fix errors in logs
Added a check so the managers will be initialized if any feature of MVCF is enabled as otherwise it'll cause issues due to the managed verbs not being registered (the patch that registers them isn't applied if no MVCF features are on).

On top of that, added extra check to only init verb managers in MP (no point touching SP).