Added compat for Performance Optimizer
Could definitely use more testing. As for testing I did: I started a game and joined with one client and then let my game run four around an hour on speed 4. No desyncs.

And most importantly: big shoutout to NotFood for pointing me to the correct method to clear the cache in.