Added compat for Performance Optimizer (#272)
Could definitely use more testing. As for testing I did: I started a game and joined with one client and then let my game run four around an hour on speed 4. No desyncs.

And most importantly: big shoutout to NotFood for pointing me to the correct method to clear the cache in.

Co-authored-by: Sokyran <toCy&#$VIH$1cSOu8s&Fv@UVN%L3snN*H09$1^Xghq*hKo9Itnmq#bNf8Ld67#T$#a&cEaY1>