Added compat for Vanilla Factions Expanded - Empire
The compat includes a "Temporary Patches" region. This includes syncing of an inner class that could not be properly synced through the use of sync delegate (pawn not accessible). The normal patch has been commented out with a // TODO note about them, mentioning the need for an API update.

This also includes a new entry inside of MpSyncWorkers for LordJob. This is definitely something that MP should handle - currently, it only does so for 2 specific ones. I'll make a PR with that at a later date.

One part of the mod was untested - discarding poisoned food. VFE-Empire seems to not include any debug methods for creating those, there seems to not be any way to spawn food with specific ingredients, and hoping for it to happen randomly was taking too long, so I've skipped it for now. It should work fine as it uses `Pawn_JobTracker.TryTakeOrderedJob`, but it could use testing out first.

The patch was made as the referenced mod, as the patch was significantly longer with a lot of harmony calls, making it much harder to maintain and read. As a small example of how it looked like - check the temporary patch (as it currently was not possible to handle it differently).