Updated VSE - Winston Waves patch
Seems some changes behind-the-scenes in the mod itself resulted in it no longer causing issues in MP and the waves generate properly now... For the most part. In my testing the waves seemed to fail to generate a few times, generally either after changing the storyteller (for 2nd wave only), or just randomly. This was always fixed by loading the game.

The patch required a couple of changes as the mod itself was modified since the patch was originally made, causing parts of the old patch obsolete or broken.

Also the button replacement patch required a slight update to support 1.4 properly.