Patched special terrain from Vanilla Expanded Framework
The terrain was using real-life time and speed of the PC that was running to end processing the terrain early if it was taking too long.

For obvious reasons, this causes with MP - if a terrain tile for example pushes heat when ticking (like lava from ReGrowth: Core) - then if it only happens for one person would be bad.

There are also situations where terrain ticking would not be problematic - things like non-ticking comps, or comps that only change the visuals (potentially mote-spawning comps).

On the negative side - this could potentially lower the performance if there's a lot of special terrains. Sadly, this is a price we'll have to pay to remove desyncs related to special terrain.