Fixed weather overlay effects in Vanilla Expanded Framework
The overlay stores a field to when next to damage calculation. The issue with this is that it's only manipulated during ticking, and never exposed at all (`SkyOverlay` is not exposable).

This also made me realize a bug with this class - the code will only ever affect a single map if there's multiple active with this same overlay, as one `SkyOverlay` will be shared for the same weather across all maps.