General maintenance to Vanilla Expanded mod patches (#321)
Vanilla Furniture Expanded - Security got cleaned quite a bit. The patch itself was quite old, and a lot of the stuff it did was unnecessary.

In Vanilla Events Expanded, the option to remove map/world condition in settings got synced. This required a sync worker for `GameCondition`, it'll need to be included in MP as well.

Vanilla Events Expanded and Vanilla Psycasts Expanded now include patches to remove the current map usage.

A few other patches that were no longer needed got removed.

Co-authored-by: Sokyran <toCy&#$VIH$1cSOu8s&Fv@UVN%L3snN*H09$1^Xghq*hKo9Itnmq#bNf8Ld67#T$#a&cEaY1>