Another attempt at patching Performance Optimizer (#369)
* Another attempt at patching Performance Optimizer

The patch was possibly thanks to Mutation Tracing tool released by Zetrith recently. Big thanks for the tool.

The patch should work, but I don't guarantee it 100%. It may turn out there's still something else that causes issues with Multiplayer that I could not find. But for now, let's hope for the best.

Changes/additions:
- The patch got split into regions
- Added a patch that makes the mod use separate cache for interface calls
- Added an error message time control hotkeys patch fails
- Fixed autosave patch by targeting the proper MP method (different name than earlier)

* Added warning and null check to dictionary replacement code