Compat for Dubs Central Heating and Dubs Bad Hygiene Lite
Bad Hygiene Lite uses the same assembly (although potentially a bit outdated?) as Bad Hygiene. It doesn't have everything that the non-lite version does, but the code is (basically) the same.

Central Heating assembly is a stripped-down version of Bad Hygiene assembly with the namespace being renamed as well. It still has all the `[SyncMethod]` attributes where those are needed, but it never call `MP.RegisterAll()` (as the class it was done it ended up being removed). Like Bad Hygiene, it also needs a sync worker for the deconstruct pipe designator.