Fixed gizmo syncing for CompGlowerExtended using wall lights from VEF
Specifically, this fixes syncing of vanilla gizmos for lights, like changing light color, switching them to darklights, or copy/pasting a color.

The exact explanation of what the issue was should be included in the comments in the code.

This should fix wall lights from Vanilla Factions Expanded - Ancients. (There may be more mods this fixes.)