Added compat for JecsTools (#382)
This has been lingering for a while, as I needed it for a compat I never finished. I've finally decided to clean it up and publish it.

It contains compat for all components (I believe) needed compat. I was only able to test `AbilityUser` component, as none of the mods I had installed user the other components of JecsTools.