Updated CashRegister compat
It's been laying around for a couple of months untouched, waiting for MP API update so I could update it alongside Gastronomy.

It's now a referenced mod, as Gastronomy was referencing it as well and it was causing issues not having this mod referenced as well - so I've decided to fully modify the patch to take advantage of that.

On top of that, there's also many fixes compared to previous version, as well as a fix for re-cached data based on real time usage.