Updated Vanilla Vehicles Expanded/added Tier 3 support
Anything from VVE Tier 3 that needed patching was included in the original vehicle mod (despite it not using it), so it acts as a framework. No need for separate patch for Tier 3 assembly, as it barely does anything (and nothing that needs patching).

Added temporary sync workers for Designation and DesignationManager, simplifying existing patch and one of the new ones.