Fix issue with VEF expandable projectiles (#395)
There was an issue when an expandable projectile was used by a moving pawn, this should fix it. The issue was that the starting position and cached value for if pawn moved was recalculated in interface. The fix is to simply discard those values in interface, restoring the ones before the call happened.