Added compat for XVI-MECHFRAME
One thing I didn't realize when starting the work on this compat was that (apparently) this mod has a bit of issues (with 1.4? or mod conflicts?)... I decided there's no point in deleting it once it was 90% done, and that I may as well finish it. However, be aware that there may some issues with the mod itself.

There's 2 versions on workshop, the original (2844617885) and the unofficial 1.4 release (3108258542). The original is on 1.4 now as well, so that's the release I tested everything with. Although there shouldn't be much of a difference, I've only tested the official release so - use the unofficial one at own risk.

The mod includes separate .dll file for both vanilla and Combat Extended. They seem very similar on the first glance and CE seemed to work fine, however - I've only tested the non-CE version. Use CE on your own risk.