Partial compat for Vanilla Factions Expanded - Tribals
This should handle the majority of the mod. However, there are a few problematic places that aren't fixed. I currently have no plans in the near future to fix them either due to complexity, or due to no API to work with async time/multifaction.

Firstly, due to the complexity the creation of player ideology after researching culture is not synced. Patching it would require making a subtype of `Page_ConfigureIdeo` and `Page_ConfigureFluidIdeo` that would handle creation and selection of an ideology when playing (harmony patches would be possible, but it would require patching significantly more vanilla methods to work). Still, those 2 types have quite a bit of stuff that could cause issues in MP, like creating and modifying ideologies in interface code, some of which would require extra patches.

However, this interaction (ignoring multifaction, which would not work properly anyway - not like anyone but the host could pick the specific start in which this is relevant) will only be used once per a save game - so a safe workaround would be for the host to create the ideology (or pick classic mode) and then create a joinpoint, which will allow others to safely play.

Another issue is with multifaction. The game component works on a bit of data, like player tech level or cornerstones, which will be shared across all players. There's currently quite a bit possibilities for stuff to break, and there's at the moment not going to be any patches to resolve this.

Last (and the smallest) issue would be with async time. The big fire building has a timer. The timer is shared for all of them (in GameComp) and the fire is processed once per tick (checks if the saved tick is different from current one). One issue is that in async time it will tick per each map (so it'll result in faster progress there). The second issue is that once every 600000 ticks it spawns a wild person on a random map with active large fire - which could target paused maps in async.