Added `MethodOf` methods, changed C# lang version to 12 (#421)
* Added 2 `MethodOf` methods, changed C# lang version to 12

The first `MethodOf` is taken from Fishery (but capitalized, as opposed to `methodof` in Fishery, and doesn't include global using). Link to the source provided in comments. **Do we need to include the copyright notice with it as well? (Mozilla Public License, v. 2.0)**

C# lang version needs to be 10+ for simple usage of Fishery's `MethodOf`, as otherwise it'll require typing more code, lowering the convenience of it.

The second `MethodOf` is taken from Microsoft Bot Builder SDK V3 (under MIT license). It's less convenient to use, but supports non-static methods without any issues.

Comparison between the results of those 2 methods: https://dotnetfiddle.net/Cmt774

I've included those methods in `MpMethodUtil`, as the class seemed fitting in name. I've also added regions to the file, splitting to lambdas part and MethodOf part.

* Dropped the alternative MethodOf

It was accessible through Harmony. Changed the comments to point to it as an alternative for non-static method, but also added a warning about potential issues.