Fix and improve MpCompatPatchLoader.LoadPatch (#422)
The `LoadPatch` method will now check all declared methods and fields in a type, including non-static ones.

`MpCompatSyncMethodAttribute` doesn't have to be applied to a non-static methods, and will now work properly when applied to them.

`MpCompatPatchAttribute` (and its subtypes) and `MpCompatSyncWorkerAttribute` need to be static - so they'll now throw an exception if they happen to be applied to a non-static method. This could help find mistakes made while developing patches.

`MpCompatSyncFieldAttribute` needs to be static and `ISyncField` must be assignable to it - an exception will be thrown when that's not the case. This change was done to both match the new behaviour of method patching, as well as to find any potential mistakes made while developing patches.