Added compat for Vanilla Temperature Expanded (#436)
The compat required a new sync worker for Vanilla Expanded Framework (for `PipeNet`) - specifically, the link/unlink interaction. It could technically be achieved differently (don't sync the pipe net but get it from the `Thing`), but this was the simplest solution and could come in handy if we ever need to sync `PipeNet` again in the future.

Likely the last, or one of the last 1.4 PRs from me.