Updated Vanilla Expanded Framework outpost compat
The rename dialogs were changed overall in 1.5, so the current handling no longer works. The syncing of renaming should be possible to handle in a somewhat universal matter by MP itself.

Syncing of removing a pawn from an outpost and creating a caravan with them was simplified by using `TransformField` call.

All the prefixes replacing Outpost methods with MP syncable ones will now execute the original method when out of MP.

Equipping apparel/weapons/other equipment from WITab was synced. It required a few workaround, but should be fine now.