Update and fix cache clearing, fixes #446
`Multiplayer.Client.ImplSerialization:Init` caches different (and more) stuff now. The data that we needed to re-cache (list of ISyncSimple implementations and since recently Session subclasses) is now handled in `Multiplayer.Client.ApiSerialization`.

On top of that, I've also added clearing of `subClassesOrdered` dictionary in `Multiplayer.Client.Util.TypeCache` to clear all the dictionaries before re-caching (as it was the only one that wasn't cleared, which I assume was due to it being added to MP after the cache clearing was made here).