Fix RimFridge warnings (#457)
Changed current map usage patches to accept another argument (`logIfMissingMethod`) which, if true, will stop warnings about missing methods.

Changed RimFridge to set this argument to true, stopping warning. The patch is kept for RimFridge forks that haven't fixed this bug (currently the only 1.5 fork has this issue fixed, but there may be a new ones in the future).