Compat for Vanilla Factions Expanded - Insectoids 2 (#458)
This should handle everything the mod has to offer, although I may have missed something. Let me know if something doesn't work.

I've included this as a separate compat since it's a separate mod (sequel), with a separate ID and mod page. The original is still left where it was. Considering it's unlikely that it will be updated, should we delete the old compat?