Update Vehicle Framework to sync designators, remove redundant patch (#460)
First, the patch to remove red cells being drawn in some specific situations. This was fixed in the mod itself, so the patch is now redundant.

Second, 2 patches and syncing was added to road area designators. The mod has 2 designators (add/remove area), and the add designator has 2 modes (prioritize/avoid, stored as a static field). Changing the mode was not synced previously, so this change should handle it:
- The road mode (prioritize/avoid) is stored by MP Compat whenever the player changes it
- Whenever the player interacts with the designator, their selected value is restored from the value stored by MP Compat
- When syncing the designator, that player's value will be synced and applied to any interaction that will use it