Added compat for Vanilla Aspirations Expanded (#461)
* Added compat for Vanilla Aspirations Expanded

The mod uses vanilla growth moment letter and dialog as its base for 2 of its own letters/dialogs.

In case of "Aspirations Fulfilled" letter/dialog, those are basically the vanilla growth moment letter/dialog with a slight changes. It's mostly handled by MP syncing, only needing the default choice handling (for which we use the vanilla growth moment default choice method).

For "Growth Moment with Aspirations" dialog (which replaces the vanilla one when a pawn reaches its last growth moment), this letter and dialog (due to extra features and no easy way to modify vanilla dialog itself) include a lot of new code, as well as copying a big chunk of the vanilla dialog's (with some changes).

The second letter required extra work to properly handle. We needed to seed the aspiration choices (similarly to how MP handles vanilla's trait and passion choices). We also need to prevent the removal of the letter when the dialog is finished, similarly to how MP does it. And finally, we needed the default choice handling for aspiration selection as well (although we use MP method to handle traits/passions).

* Include GitHub/Workshop links