Update Designator Shapes to add a hard 10000 cells limit in MP (#463)
Added a patch to Designator Shapes flood fill code, which will temporarily set the maximum amount of cells (when in a Multiplayer session) to 10000 (if the actual setting is above that value).

When trying to designate over (roughly) 10900 cells, MP will fail and cause an exception due to packet size being too big. On top of that, if the packet size is in a very close to the packet size limit it may cause a remote packet read error for the clients (effectively kicking them).

This is an issue that can also happen with vanilla designators (among other things), but that is something that should be fixed on MP side. The most we can do in MP Compat is to try and mitigate it in mods that may encounter this issue, which this is trying to achieve. So until this is fixed by MP, it may be a good idea to have this little workaround here.