Added compat for UI Not Included
Well, I may have overdone it on the error checking part a bit... However, I've decided that if we patch something and that patch only partially works, it may make the game unplayable. So way too much safety over potential issues in the future.

This compat fixes:
- The mod's custom time control widget using vanilla time control UI, rather than the one introduced by MP
- Time controls hotkeys not working, unless using vanilla time control UI in options

The fix is done by:
- Replacing the cached method `Action<Rect>` the mod was using to draw the vanilla/Smart Speed's speed control UI
  - This will also handle the hotkeys, as long as the widget is in use
- Calling the Multiplayer method handling hotkeys
  - This will only happen if both vanilla speed control and the mod's custom speed control widget are off

One unpatched thing is that the "play settings at top" option will cause the play settings buttons to be covered by Multiplayer chat and (if enabled) debug buttons/data. It's not as game breaking as inability to unpause/change speed, so I didn't bother fixing it right now.