Fix Vehicle Framework issues with upgrade overlays
When loading a game where a vehicle has an upgrade that applies an overlay to a vehicle, the overlay initialization would fail for the host causing issues and desyncs.

This should fix all of those issues by patching the method and forcing it to run inside a `LongEventHandler.ExecuteWhenFinished` call, delaying it until it's safe to run it.