Fix wrong target for Harmony prefix for Vanilla Recycling Expanded
It looks like I've targeted a wrong method with a Harmony prefix. This caused 2 issues:

The first is that the prefix caused reclaimed biopacks from properly dissolving, throwing an exception and potentially causing quite a bit of lag.

The second is that the method that was meant to be patched could be called when it's not supposed to, meaning that the item processor would attempt to reduce the stack size by 1 while it's at 0. Only possible in dev mode, but still possible.

This should change the method to patch the correct method. On top of that, I've updated 1 of the comments in there as well.