Fix ReGrowth: Core caravan camp maps not being removed
This happens due to one of MP patches that disables automatic removal of world object maps if the world object is owned by any faction. The patch I've applied here simply does the same check that is disabled by MP, but rather than remove the map it just sets the faction to null (allowing the map to be removed naturally).