Update PatchingUtilities.ReplaceMethod method
The updated method accepts 2 `Func<CodeInstruction, bool>` arguments, which are used to determine if the next method should or should not be patched.

Those are mutually exclusive with the string `targetText`/`excludedText`. They'll now create a function that will check if the instruction loads a string and that string matches their target.