Add debug action to use "Use This Instead" mod to search for outdated compats (#526)
Second attempt at pushing those changes, as the first time it ended up with the whole file being seen as modified.

Basically, a new debug action was added that will use "Use This Instead" mod to search its database, and compare it with all the compats we have made. It will log every single mod ID that has replacements, the replacements ID, and replacements workshop URL (unless we also support the replacement as well, in which case it will be skipped).

This was done over a span of roughly 30 minutes at 2 AM, so it's not perfect. Specifically, it only searches loaded MP Compat assemblies - meaning that the referenced mods won't be checked unless they're also active while we're searching.